Update technical-charter.md to remove TSC voter eligibility requirements (#96)
The charter contains language that requires members to have attended TSC meetings frequently to be eligible to vote. This requirement is not common in other open source project charters, and it turns out, makes it very difficult to have an electronic vote.

1. Tracking eligibility requirements is cumbersome
2. When performing an electronic vote, it is even more cumbersome to keep track of who is eligible and who is not
3. If during the time limit for the vote a member may become eligible, or may become ineligible, does their vote count?

I propose we remove this requirement. We will then be able to do electronic / asynchronous votes more clearly and easily. We have learned from experience thus far that gaining quorum in the TSC meetings has been difficult, making it challenging to get votes completed. By removing this restriction, we should be able to move forward the work faster.

Signed-off-by: Richard Bair <3486721+rbair23@users.noreply.github.com>